+++ node_helper.js
@@ -15,42 +15,7 @@ module.exports = NodeHelper.create({
 	// Create volvoApiClient object
 	volvoApiClient: null,
 
-	start: function () {
-		// Setup the local endpoints needed for the OAuth2
-		this.expressApp.get(`/${this.name}/auth`, this.loginHandler.bind(this));
-		this.expressApp.get(`/${this.name}/callback`, this.callbackHandler.bind(this));
-	},
-
-
-	loginHandler: function (req, res) {
-		// Get the fullAuthUrl + csrfState parameters from the authClient
-		var [fullAuthUrl, csrfState] = this.authClient.getAuthorizationParameters();
-
-		// Set a cookie with the csrfState to be able to verify later
-		res.cookie('MMM-VolvoCar-state', csrfState, { maxAge: 1000 * 60 * 5 });
-
-		// Redirect to the fullAuthUrl
-		res.redirect(fullAuthUrl);
-	},
-
-	callbackHandler: function (req, res) {
-		// Get the authorization code from the callback
-		var { code, state } = req.query;
-
-		// Parse the cookies to get the state cookie set in the previous auth request
-		var { groups: { csrfState } } = /(MMM-VolvoCar-state)=(?<csrfState>.*?)($|;|,(?! ))/.exec(req.headers.cookie)
-
-		// Verify that the state value we got back is the same as we previously sent, otherwise abort
-		if (state !== csrfState) {
-			res.status(422).send('Invalid state')
-			return;
-		}
-
-		// Get the access token using the authorization code
-		this.authClient.getToken(code, () => {
-			res.redirect('/');
-		});
-	},
+	start: function () {},
 
 	downloadHeaderImage: function (imageUrl, localPath, carData) {
 		var self = this;
@@ -101,8 +66,8 @@ module.exports = NodeHelper.create({
 			}
 
 			// Make sure that we have needed credential set in the config
-			if (!this.config.authClientId || !this.config.authClientSecret || !this.config.authVccApiKey || !this.config.carVin || !this.config.carType) {
-				Log.error(this.name + ' - MMMVC_SET_CONFIG: Either authClientId, authClientSecret, authVccApiKey, carVin or carType is not set, exiting...');
+			if (!this.config.authUsername || !this.config.authPassword || !this.config.authVccApiKey || !this.config.carVin || !this.config.carType) {
+				Log.error(this.name + ' - MMMVC_SET_CONFIG: Either authUsername, authPassword, authVccApiKey, carVin or carType is not set, exiting...');
 				return;
 			}
 
@@ -121,7 +86,7 @@ module.exports = NodeHelper.create({
 			// Check if an access token exists and if it is valid or not
 			if (this.authClient.isExpired()) {
 				if (this.authClient.refresh_token && this.authClient.refresh_token.length > 0) {
-					this.authClient.refreshToken(() => {
+					this.authClient.refreshToken(this.authClient.refresh_token, () => {
 						self.sendSocketNotification('MMMVC_MODULE_READY');
 					});
 				}
@@ -137,7 +102,7 @@ module.exports = NodeHelper.create({
 			// Make sure that the authClient is setup and access_token is valid
 			if (this.authClient === null) return;
 			if (this.authClient.isExpired()) {
-				this.authClient.refreshToken(() => {
+				this.authClient.refreshToken(this.authClient.refresh_token, () => {
 					self.sendSocketNotification('MMMVC_GET_CAR_DATA');
 				});
 				return;
@@ -162,6 +127,9 @@ module.exports = NodeHelper.create({
 			Promise.all([
 				// Energy API
 				this.volvoApiClient.getRechargeStatus(this.authClient.access_token),
+
+				// Location API
+				this.volvoApiClient.getCarLocation(this.authClient.access_token),
 
 				// Connected Vehicle API
 				this.volvoApiClient.getVehicleDetails(this.authClient.access_token),