+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [2.0.0] - 2023-07-31
+
+### Added
+
+- Added support for the new Locations API
+
+### Changed
+
+- Changed OAuth 2.0 to the password flow
+- Some minor adjustments on the layout
+
+
 ## [1.2.0] - 2023-03-30
 
 ### Added